hwmon: ltc2947: Support GPIO configuration
The LTC2947 has a GPIO pin which can be configured as input or output.
When configured as input, it can be used to control the accumulation of
charge, energy and time (this driver only supports energy). This
function can be enabled/configured separately for each of the two sets
of accumulation registers.
When configured as output, it can be used to control a fan as a function
of measured temperature by setting the corresponding thresholds. For
instance, the pin becomes active as soon as the temperature measurement
is above the threshold TFANH.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>